@@ -16,27 +16,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.0",
+    "@angular/animations": "~10.2.3",
     "@angular/cdk": "^9.0.0",
-    "@angular/common": "~9.1.0",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "~9.1.0",
-    "@angular/forms": "~9.1.0",
+    "@angular/common": "~10.2.3",
+    "@angular/compiler": "~10.2.3",
+    "@angular/core": "~10.2.3",
+    "@angular/forms": "~10.2.3",
     "@angular/material": "^9.0.0",
-    "@angular/platform-browser": "~9.1.0",
-    "@angular/platform-browser-dynamic": "~9.1.0",
-    "@angular/router": "~9.1.0",
+    "@angular/platform-browser": "~10.2.3",
+    "@angular/platform-browser-dynamic": "~10.2.3",
+    "@angular/router": "~10.2.3",
     "@briebug/jest-schematic": "^2.1.1",
     "@schuchard/prettier": "^3.1.0",
     "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.0",
-    "@angular/cli": "~9.1.0",
-    "@angular/compiler-cli": "~9.1.0",
-    "@angular/language-service": "~9.1.0",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "~10.2.0",
+    "@angular/compiler-cli": "~10.2.3",
+    "@angular/language-service": "~10.2.3",
     "@ngneat/spectator": "^5.3.1",
     "@types/jest": "25.1.4",
     "@types/node": "^12.11.1",
@@ -48,12 +48,12 @@
     "lint-staged": "10.0.9",
     "prettier": "2.0.2",
     "pretty-quick": "^2.0.1",
-    "protractor": "~5.4.3",
+    "protractor": "~7.0.0",
     "sass-lint": "^1.13.1",
     "ts-node": "~8.8.1",
     "tslint": "~6.1.0",
     "tslint-config-prettier": "1.18.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.0.5"
   },
   "jest": {
     "preset": "jest-preset-angular",
