@@ -10,7 +10,9 @@ import { StorageService } from './services/storage.service';
   providers: [StorageService]
 })
 export class StorageModule {
-  static withOptions(options: { storagePrefix: string }): ModuleWithProviders {
+  static withOptions(options: {
+    storagePrefix: string;
+  }): ModuleWithProviders<StorageModule> {
     return {
       ngModule: StorageModule,
       providers: [
